Update FFI definitions for pylifecycle.h
